+++ src/image/mod.rs
@@ -7,21 +7,16 @@ pub mod simple;
 pub mod rgba;
 
 use crate::meta::attributes::*;
-use crate::compression::{Compression, ByteVec};
-use crate::math::*;
-use std::io::{Read, Seek, Write, Cursor};
-use crate::error::{Result, Error, UnitResult, usize_to_i32};
-use crate::meta::{MetaData, Header, TileIndices, Blocks};
-use crate::chunks::{Chunk, Block, TileBlock, ScanLineBlock, TileCoordinates};
+use crate::compression::{Compression};
+use std::io::{Read, Seek};
+use crate::error::{Result, UnitResult};
+use crate::meta::{MetaData, Header, TileIndices};
+use crate::chunk::{Chunk};
 use crate::io::{PeekRead, Tracking};
 use rayon::iter::{ParallelIterator, ParallelBridge};
-use crate::io::Data;
-use smallvec::SmallVec;
-use std::ops::Range;
 use std::convert::TryFrom;
 use std::collections::BTreeMap;
-
-
+use crate::block::{BlockIndex, UncompressedBlock};
 
 
 /// Specify how to write an exr image.
@@ -131,90 +126,6 @@ pub mod read_options {
 }
 
 
-/// Specifies where a block of pixel data should be placed in the actual image.
-/// This is a globally unique identifier which
-/// includes the layer, level index, and pixel location.
-#[derive(Clone, Copy, Eq, Hash, PartialEq, Debug)]
-pub struct BlockIndex {
-
-    /// Index of the layer.
-    pub layer: usize,
-
-    /// Pixel position of the bottom left corner of the block.
-    pub pixel_position: Vec2<usize>,
-
-    /// Pixel size of the block.
-    pub pixel_size: Vec2<usize>,
-
-    /// Index of the mip or rip level in the image.
-    pub level: Vec2<usize>,
-}
-
-/// Contains a block of pixel data and where that data should be placed in the actual image.
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub struct UncompressedBlock {
-
-    /// Location of the data inside the image.
-    pub index: BlockIndex,
-
-    /// Uncompressed pixel values of the whole block.
-    /// One or more scan lines may be stored together as a scan line block.
-    /// This byte vector contains all pixel rows, one after another.
-    /// For each line in the tile, for each channel, the row values are contiguous.
-    pub data: ByteVec,
-}
-
-/// A single line of pixels.
-/// Use `LineRef` or `LineRefMut` for easier type names.
-#[derive(Clone, Copy, Eq, PartialEq, Debug)]
-pub struct LineSlice<T> {
-
-    /// Where this line is located inside the image.
-    pub location: LineIndex,
-
-    /// The raw bytes of the pixel line, either `&[u8]` or `&mut [u8]`.
-    /// Must be re-interpreted as slice of f16, f32, or u32,
-    /// according to the channel data type.
-    pub value: T,
-}
-
-
-/// An reference to a single line of pixels.
-/// May go across the whole image or just a tile section of it.
-///
-/// This line contains an immutable slice that all samples will be read from.
-pub type LineRef<'s> = LineSlice<&'s [u8]>;
-
-/// A reference to a single mutable line of pixels.
-/// May go across the whole image or just a tile section of it.
-///
-/// This line contains a mutable slice that all samples will be written to.
-pub type LineRefMut<'s> = LineSlice<&'s mut [u8]>;
-
-
-/// Specifies where a row of pixels lies inside an image.
-/// This is a globally unique identifier which includes
-/// the layer, channel index, and pixel location.
-#[derive(Clone, Copy, Eq, PartialEq, Debug, Hash)]
-pub struct LineIndex {
-
-    /// Index of the layer.
-    pub layer: usize,
-
-    /// The channel index of the layer.
-    pub channel: usize,
-
-    /// Index of the mip or rip level in the image.
-    pub level: Vec2<usize>,
-
-    /// Position of the most left pixel of the row.
-    pub position: Vec2<usize>,
-
-    /// The width of the line; the number of samples in this row,
-    /// that is, the number of f16, f32, or u32 values.
-    pub sample_count: usize,
-}
-
 /// Called occasionally when writing a file.
 /// Implemented by any closure that matches `|progress: f32, bytes_written: usize| -> UnitResult`.
 pub trait OnWriteProgress {
@@ -252,91 +163,14 @@ impl OnReadProgress for () {
 }
 
 
-impl<'s> LineRefMut<'s> {
 
-    /// Writes the samples (f16, f32, u32 values) into this line value reference.
-    /// Use `write_samples` if there is not slice available.
-    #[inline]
-    #[must_use]
-    pub fn write_samples_from_slice<T: crate::io::Data>(self, slice: &[T]) -> UnitResult {
-        debug_assert_eq!(slice.len(), self.location.sample_count, "slice size does not match the line width");
-        debug_assert_eq!(self.value.len(), self.location.sample_count * T::BYTE_SIZE, "sample type size does not match line byte size");
-
-        T::write_slice(&mut Cursor::new(self.value), slice)
-    }
-
-    /// Iterate over all samples in this line, from left to right.
-    /// The supplied `get_line` function returns the sample value
-    /// for a given sample index within the line,
-    /// which starts at zero for each individual line.
-    /// Use `write_samples_from_slice` if you already have a slice of samples.
-    #[inline]
-    #[must_use]
-    pub fn write_samples<T: crate::io::Data>(self, mut get_sample: impl FnMut(usize) -> T) -> UnitResult {
-        debug_assert_eq!(self.value.len(), self.location.sample_count * T::BYTE_SIZE, "sample type size does not match line byte size");
-
-        let mut write = Cursor::new(self.value);
-
-        for index in 0..self.location.sample_count {
-            T::write(get_sample(index), &mut write)?;
-        }
-
-        Ok(())
-    }
-}
-
-impl LineRef<'_> {
-
-    /// Read the samples (f16, f32, u32 values) from this line value reference.
-    /// Use `read_samples` if there is not slice available.
-    pub fn read_samples_into_slice<T: crate::io::Data>(self, slice: &mut [T]) -> UnitResult {
-        debug_assert_eq!(slice.len(), self.location.sample_count, "slice size does not match the line width");
-        debug_assert_eq!(self.value.len(), self.location.sample_count * T::BYTE_SIZE, "sample type size does not match line byte size");
-
-        T::read_slice(&mut Cursor::new(self.value), slice)
-    }
-
-    /// Iterate over all samples in this line, from left to right.
-    /// Use `read_sample_into_slice` if you already have a slice of samples.
-    pub fn read_samples<T: crate::io::Data>(&self) -> impl Iterator<Item = Result<T>> + '_ {
-        debug_assert_eq!(self.value.len(), self.location.sample_count * T::BYTE_SIZE, "sample type size does not match line byte size");
-
-        let mut read = self.value.clone(); // FIXME deep data
-        (0..self.location.sample_count).map(move |_| T::read(&mut read))
-    }
-}
 
 
 /// Reads and decompresses all chunks of a file sequentially without seeking.
 /// Will not skip any parts of the file. Does not buffer the reader, you should always pass a `BufReader`.
 #[inline]
 #[must_use]
-pub fn read_all_lines_from_buffered<T>(
-    read: impl Read + Send, // FIXME does not actually need to be send, only for parallel writing
-    new: impl Fn(&[Header]) -> Result<T>,
-    mut insert: impl FnMut(&mut T, &[Header], LineRef<'_>) -> UnitResult,
-    options: ReadOptions<impl OnReadProgress>,
-) -> Result<T>
-{
-    let insert = |value: &mut T, headers: &[Header], decompressed: UncompressedBlock| {
-        let header = headers.get(decompressed.index.layer)
-            .ok_or(Error::invalid("chunk index"))?;
-
-        for (bytes, line) in decompressed.index.line_indices(header) {
-            insert(value, headers, LineSlice { location: line, value: &decompressed.data[bytes] })?; // allows returning `Error::Abort`
-        }
-
-        Ok(())
-    };
-
-    read_all_tiles_from_buffered(read, new, insert, options)
-}
-
-/// Reads and decompresses all chunks of a file sequentially without seeking.
-/// Will not skip any parts of the file. Does not buffer the reader, you should always pass a `BufReader`.
-#[inline]
-#[must_use]
-pub fn read_all_tiles_from_buffered<T>(
+pub fn read_all_blocks_from_buffered<T>(
     read: impl Read + Send, // FIXME does not actually need to be send, only for parallel writing
     new: impl Fn(&[Header]) -> Result<T>,
     mut insert: impl FnMut(&mut T, &[Header], UncompressedBlock) -> UnitResult,
@@ -349,7 +183,7 @@ pub fn read_all_tiles_from_buffered<T>(
     let read_chunks = std::iter::from_fn(move || read_chunk(meta_data_ref));
     let mut result = new(meta_data.headers.as_slice())?;
 
-    for_decompressed_tiles_in_chunks(
+    for_decompressed_blocks_in_chunks(
         read_chunks, &meta_data,
         |meta, block| insert(&mut result, meta, block),
         chunk_count, options
@@ -359,34 +193,6 @@ pub fn read_all_tiles_from_buffered<T>(
 }
 
 
-/// Reads and decompresses all desired chunks of a file sequentially, possibly seeking.
-/// Will skip any parts of the file that do not match the specified filter condition.
-/// Will never seek if the filter condition matches all chunks.
-/// Does not buffer the reader, you should always pass a `BufReader`.
-#[inline]
-#[must_use]
-pub fn read_filtered_lines_from_buffered<T>(
-    read: impl Read + Seek + Send, // FIXME does not always need be Send
-    new: impl Fn(&[Header]) -> Result<T>, // TODO put these into a trait?
-    filter: impl Fn(&T, &Header, &TileIndices) -> bool,
-    mut insert: impl FnMut(&mut T, &[Header], LineRef<'_>) -> UnitResult,
-    options: ReadOptions<impl OnReadProgress>,
-) -> Result<T>
-{
-    let insert = |value: &mut T, headers: &[Header], decompressed: UncompressedBlock| {
-        let header = headers.get(decompressed.index.layer)
-            .ok_or(Error::invalid("chunk index"))?;
-
-        for (bytes, line) in decompressed.index.line_indices(header) {
-            insert(value, headers, LineSlice { location: line, value: &decompressed.data[bytes] })?; // allows returning `Error::Abort`
-        }
-
-        Ok(())
-    };
-
-    read_filtered_tiles_from_buffered(read, new, filter, insert, options)
-}
-
 
 /// Reads ad decompresses all desired chunks of a file sequentially, possibly seeking.
 /// Will skip any parts of the file that do not match the specified filter condition.
@@ -394,7 +200,7 @@ pub fn read_filtered_lines_from_buffered<T>(
 /// Does not buffer the reader, you should always pass a `BufReader`.
 #[inline]
 #[must_use]
-pub fn read_filtered_tiles_from_buffered<T>(
+pub fn read_filtered_blocks_from_buffered<T>(
     read: impl Read + Seek + Send, // FIXME does not always need be Send
     new: impl FnOnce(&[Header]) -> Result<T>, // TODO put these into a trait?
     filter: impl Fn(&T, &Header, &TileIndices) -> bool,
@@ -406,7 +212,7 @@ pub fn read_filtered_tiles_from_buffered<T>(
         self::read_filtered_chunks_from_buffered(read, new, filter, options.max_pixel_bytes)?
     };
 
-    for_decompressed_tiles_in_chunks(
+    for_decompressed_blocks_in_chunks(
         std::iter::from_fn(|| read_chunk(&meta_data)), &meta_data,
         |meta, line| insert(&mut value, meta, line),
         chunk_count, options
@@ -419,7 +225,7 @@ pub fn read_filtered_tiles_from_buffered<T>(
 /// Decompresses the chunks either in parallel or sequentially.
 #[inline]
 #[must_use]
-fn for_decompressed_tiles_in_chunks(
+fn for_decompressed_blocks_in_chunks(
     chunks: impl Send + Iterator<Item = Result<Chunk>>,
     meta_data: &MetaData,
     mut for_each: impl FnMut(&[Header], UncompressedBlock) -> UnitResult,
@@ -541,7 +347,7 @@ pub fn read_filtered_chunks_from_buffered<'m, T>(
 #[must_use]
 pub fn uncompressed_image_blocks_ordered<'l>(
     meta_data: &'l MetaData,
-    get_tile: &'l (impl 'l + Sync + (Fn(&[Header], BlockIndex) -> Vec<u8>)) // TODO reduce sync requirements, at least if parrallel is false
+    get_block: &'l (impl 'l + Sync + (Fn(&[Header], BlockIndex) -> Vec<u8>)) // TODO reduce sync requirements, at least if parrallel is false
 ) -> impl 'l + Iterator<Item = Result<(usize, UncompressedBlock)>> + Send // TODO reduce sync requirements, at least if parrallel is false
 {
     meta_data.headers.iter().enumerate()
@@ -555,7 +361,7 @@ pub fn uncompressed_image_blocks_ordered<'l>(
                     pixel_size: data_indices.size,
                 };
 
-                let block_bytes = get_tile(meta_data.headers.as_slice(), block_indices);
+                let block_bytes = get_block(meta_data.headers.as_slice(), block_indices);
 
                 // byte length is validated in block::compress_to_chunk
                 Ok((chunk_index, UncompressedBlock {
@@ -650,261 +456,3 @@ pub fn for_compressed_blocks_in_image(
 
     Ok(())
 }
-
-
-/// Compresses and writes all lines of an image described by `meta_data` and `get_line` to the writer.
-/// Flushes the writer to explicitly handle all errors.
-///
-/// Attention: Currently, using multi-core compression with `LineOrder::Increasing` or `LineOrder::Decreasing` in any header
-/// can potentially allocate large amounts of memory while writing the file. Use unspecified line order for lower memory usage.
-///
-/// Does not buffer the writer, you should always pass a `BufWriter`.
-/// If pedantic, throws errors for files that may produce errors in other exr readers.
-#[inline]
-#[must_use]
-pub fn write_all_lines_to_buffered(
-    write: impl Write + Seek, meta_data: MetaData,
-    get_line: impl Sync + Fn(&[Header], LineRefMut<'_>), // TODO put these three parameters into a trait?  // TODO why is this sync or send????
-    options: WriteOptions<impl OnWriteProgress>,
-) -> UnitResult
-{
-    let get_block = |headers: &[Header], block_index: BlockIndex| {
-        let header: &Header = &headers.get(block_index.layer).expect("invalid block index");
-
-        let bytes = block_index.pixel_size.area() * header.channels.bytes_per_pixel;
-        let mut block_bytes = vec![0_u8; bytes];
-
-        for (byte_range, line_index) in block_index.line_indices(header) {
-            get_line(headers, LineRefMut {
-                value: &mut block_bytes[byte_range],
-                location: line_index,
-            });
-        }
-
-        block_bytes
-    };
-
-    write_all_tiles_to_buffered(write, meta_data, get_block, options)
-}
-
-/// Compresses and writes all lines of an image described by `meta_data` and `get_line` to the writer.
-/// Flushes the writer to explicitly handle all errors.
-///
-/// Attention: Currently, using multi-core compression with `LineOrder::Increasing` or `LineOrder::Decreasing` in any header
-/// can potentially allocate large amounts of memory while writing the file. Use unspecified line order for lower memory usage.
-///
-/// Does not buffer the writer, you should always pass a `BufWriter`.
-/// If pedantic, throws errors for files that may produce errors in other exr readers.
-#[inline]
-#[must_use]
-pub fn write_all_tiles_to_buffered(
-    write: impl Write + Seek,
-    mut meta_data: MetaData,
-    get_tile: impl Sync + Fn(&[Header], BlockIndex) -> Vec<u8>, // TODO put these three parameters into a trait?  // TODO why is this sync or send????
-    mut options: WriteOptions<impl OnWriteProgress>,
-) -> UnitResult
-{
-    let has_compression = meta_data.headers.iter() // TODO cache this in MetaData.has_compression?
-        .any(|header| header.compression != Compression::Uncompressed);
-
-    // if non-parallel compression, we always use increasing order anyways
-    if !options.parallel_compression || !has_compression {
-        for header in &mut meta_data.headers {
-            if header.line_order == LineOrder::Unspecified {
-                header.line_order = LineOrder::Increasing;
-            }
-        }
-    }
-
-    let mut write = Tracking::new(write);
-    meta_data.write_validating_to_buffered(&mut write, options.pedantic)?; // also validates meta data
-
-    let offset_table_start_byte = write.byte_position();
-
-    // skip offset tables for now
-    let offset_table_size: usize = meta_data.headers.iter()
-        .map(|header| header.chunk_count).sum();
-
-    write.seek_write_to(write.byte_position() + offset_table_size * std::mem::size_of::<u64>())?;
-
-    let mut offset_tables: Vec<Vec<u64>> = meta_data.headers.iter()
-        .map(|header| vec![0; header.chunk_count]).collect();
-
-    let total_chunk_count = offset_table_size as f32;
-    let mut processed_chunk_count = 0; // very simple on_progress feedback
-
-    // line order is respected in here
-    for_compressed_blocks_in_image(&meta_data, get_tile, options.parallel_compression, |chunk_index, chunk|{
-        offset_tables[chunk.layer_index][chunk_index] = write.byte_position() as u64; // safe indices from `enumerate()`
-        chunk.write(&mut write, meta_data.headers.as_slice())?;
-
-        options.on_progress.on_write_progressed(
-            processed_chunk_count as f32 / total_chunk_count, write.byte_position()
-        )?;
-
-        processed_chunk_count += 1;
-        Ok(())
-    })?;
-
-    // write all offset tables
-    write.seek_write_to(offset_table_start_byte)?;
-
-    for offset_table in offset_tables {
-        u64::write_slice(&mut write, offset_table.as_slice())?;
-    }
-
-    write.flush()?; // make sure we catch all (possibly delayed) io errors before returning
-
-    Ok(())
-}
-
-
-impl BlockIndex {
-
-    /// Iterates the lines of this block index in interleaved fashion:
-    /// For each line in this block, this iterator steps once through each channel.
-    /// This is how lines are stored in a pixel data block.
-    ///
-    /// Does not check whether `self.layer_index`, `self.level`, `self.size` and `self.position` are valid indices.__
-    // TODO be sure this cannot produce incorrect data, as this is not further checked but only handled with panics
-    #[inline]
-    #[must_use]
-    pub fn line_indices(&self, header: &Header) -> impl Iterator<Item=(Range<usize>, LineIndex)> {
-        struct LineIter {
-            layer: usize, level: Vec2<usize>, width: usize,
-            end_y: usize, x: usize, channel_sizes: SmallVec<[usize; 8]>,
-            byte: usize, channel: usize, y: usize,
-        };
-
-        // FIXME what about sub sampling??
-
-        impl Iterator for LineIter {
-            type Item = (Range<usize>, LineIndex);
-
-            fn next(&mut self) -> Option<Self::Item> {
-                if self.y < self.end_y {
-
-                    // compute return value before incrementing
-                    let byte_len = self.channel_sizes[self.channel];
-                    let return_value = (
-                        (self.byte .. self.byte + byte_len),
-                        LineIndex {
-                            channel: self.channel,
-                            layer: self.layer,
-                            level: self.level,
-                            position: Vec2(self.x, self.y),
-                            sample_count: self.width,
-                        }
-                    );
-
-                    { // increment indices
-                        self.byte += byte_len;
-                        self.channel += 1;
-
-                        if self.channel == self.channel_sizes.len() {
-                            self.channel = 0;
-                            self.y += 1;
-                        }
-                    }
-
-                    Some(return_value)
-                }
-
-                else {
-                    None
-                }
-            }
-        }
-
-        let channel_line_sizes: SmallVec<[usize; 8]> = header.channels.list.iter()
-            .map(move |channel| self.pixel_size.0 * channel.sample_type.bytes_per_sample()) // FIXME is it fewer samples per tile or just fewer tiles for sampled images???
-            .collect();
-
-        LineIter {
-            layer: self.layer,
-            level: self.level,
-            width: self.pixel_size.0,
-            x: self.pixel_position.0,
-            end_y: self.pixel_position.1 + self.pixel_size.1,
-            channel_sizes: channel_line_sizes,
-
-            byte: 0,
-            channel: 0,
-            y: self.pixel_position.1
-        }
-    }
-}
-
-impl UncompressedBlock {
-
-    /// Decompress the possibly compressed chunk and returns an `UncompressedBlock`.
-    // for uncompressed data, the ByteVec in the chunk is moved all the way
-    #[inline]
-    #[must_use]
-    pub fn decompress_chunk(chunk: Chunk, meta_data: &MetaData) -> Result<Self> {
-        let header: &Header = meta_data.headers.get(chunk.layer_index)
-            .ok_or(Error::invalid("chunk layer index"))?;
-
-        let tile_data_indices = header.get_block_data_indices(&chunk.block)?;
-        let absolute_indices = header.get_absolute_block_indices(tile_data_indices)?;
-
-        absolute_indices.validate(Some(header.data_size))?;
-
-        match chunk.block {
-            Block::Tile(TileBlock { compressed_pixels, .. }) |
-            Block::ScanLine(ScanLineBlock { compressed_pixels, .. }) => Ok(UncompressedBlock {
-                data: header.compression.decompress_image_section(header, compressed_pixels, absolute_indices)?,
-                index: BlockIndex {
-                    layer: chunk.layer_index,
-                    pixel_position: absolute_indices.position.to_usize("data indices start")?,
-                    level: tile_data_indices.level_index,
-                    pixel_size: absolute_indices.size,
-                }
-            }),
-
-            _ => return Err(Error::unsupported("deep data not supported yet"))
-        }
-    }
-
-    /// Consume this block by compressing it, returning a `Chunk`.
-    // for uncompressed data, the ByteVec in the chunk is moved all the way
-    #[inline]
-    #[must_use]
-    pub fn compress_to_chunk(self, meta_data: &MetaData) -> Result<Chunk> {
-        let UncompressedBlock { data, index } = self;
-
-        let header: &Header = meta_data.headers.get(index.layer)
-            .expect("block layer index bug");
-
-        let expected_byte_size = header.channels.bytes_per_pixel * self.index.pixel_size.area(); // TODO sampling??
-        if expected_byte_size != data.len() {
-            panic!("get_line byte size should be {} but was {}", expected_byte_size, data.len());
-        }
-
-        let compressed_data = header.compression.compress_image_section(data)?;
-
-        Ok(Chunk {
-            layer_index: index.layer,
-            block : match header.blocks {
-                Blocks::ScanLines => Block::ScanLine(ScanLineBlock {
-                    compressed_pixels: compressed_data,
-
-                    // FIXME this calculation should not be made here but elsewhere instead (in meta::header?)
-                    y_coordinate: usize_to_i32(index.pixel_position.1) + header.own_attributes.data_position.1,
-                }),
-
-                Blocks::Tiles(tiles) => Block::Tile(TileBlock {
-                    compressed_pixels: compressed_data,
-                    coordinates: TileCoordinates {
-                        level_index: index.level,
-
-                        // FIXME this calculation should not be made here but elsewhere instead (in meta::header?)
-                        tile_index: index.pixel_position / tiles.tile_size,
-                    },
-
-                }),
-            }
-        })
-    }
-}
-
